requirement.txt
LangChain
streamlit
langchain-google-genai 
google-generativeai
python-dotenv
certifi
numpy